+++ lib/data/repositories/remote_user_info_reposiotory.dart
@@ -1,6 +1,6 @@
 import 'package:mongbi_app/data/data_sources/user_info_data_source.dart';
-import 'package:mongbi_app/domain/entities/user.dart';
 import 'package:mongbi_app/domain/repositories/user_info_repository.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
 
 class RemoteUserInfoRepository implements UserInfoRepository {
   final GetUserInfoDataSource dataSource;
