+++ lib/domain/repositories/nickname_repository.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/domain/entities/user.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
 
 abstract interface class NicknameSettingRepository {
   Future<User> updateNickname(int userId, String nickname);
