+++ lib/domain/use_cases/remove_account_use_case.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/domain/repositories/account_repository.dart';
+import 'package:mongbi_app/features/auth/domain/repositories/account_repository.dart';
 
 class RemoveAccountUseCase {
   RemoveAccountUseCase(this.repository);
