+++ lib/features/auth/data/data_sources/naver_auth_data_source.dart
@@ -1,4 +1,4 @@
-import 'package:mongbi_app/data/dtos/login_response_dto.dart';
+import 'package:mongbi_app/features/auth/data/dtos/login_response_dto.dart';
 
 abstract interface class AuthDataSource {
   Future<LoginResponseDto> naverLogin(String idToken);
