+++ lib/features/auth/data/dtos/remote_auth_repository.dart
@@ -1,8 +1,8 @@
-import 'package:mongbi_app/data/data_sources/remote_apple_auth_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_kakao_auth_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_naver_auth_data_source.dart';
-import 'package:mongbi_app/domain/entities/user.dart';
 import 'package:mongbi_app/domain/repositories/auth_repository.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_apple_auth_data_source.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_kakao_auth_data_source.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_naver_auth_data_source.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
 
 class RemoteAuthRepository implements AuthRepository {
   RemoteAuthRepository({
