+++ lib/features/auth/domain/entities/user.dart
