+++ lib/features/auth/domain/use_cases/get_user_info_use_case.dart
@@ -1,5 +1,5 @@
-import 'package:mongbi_app/domain/entities/user.dart';
 import 'package:mongbi_app/domain/repositories/user_info_repository.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
 
 class GetUserInfoUseCase {
   GetUserInfoUseCase(this.repository);
