+++ lib/features/auth/domain/use_cases/login_with_naver.dart
@@ -1,5 +1,5 @@
-import 'package:mongbi_app/domain/entities/user.dart';
 import 'package:mongbi_app/domain/repositories/auth_repository.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
 
 class LoginWithNaver {
   LoginWithNaver(this.repository);
