+++ lib/features/auth/presentation/pages/social_login_page.dart
@@ -4,12 +4,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
 import 'package:mongbi_app/core/analytics/analytics_helper.dart';
 import 'package:mongbi_app/core/exceptions/auth_custom_exception.dart';
-import 'package:mongbi_app/presentation/auth/widgets/apple_login_button_widget.dart';
-import 'package:mongbi_app/presentation/auth/widgets/kakao_login_button_widget.dart';
-import 'package:mongbi_app/presentation/auth/widgets/last_login_state_weiget.dart';
-import 'package:mongbi_app/presentation/auth/widgets/mongbi_image_widget.dart';
-import 'package:mongbi_app/presentation/auth/widgets/naver_login_button_widget.dart';
-import 'package:mongbi_app/presentation/auth/widgets/text_widgets.dart';
+import 'package:mongbi_app/features/auth/presentation/pages/widgets/apple_login_button_widget.dart';
+import 'package:mongbi_app/features/auth/presentation/pages/widgets/kakao_login_button_widget.dart';
+import 'package:mongbi_app/features/auth/presentation/pages/widgets/last_login_state_weiget.dart';
+import 'package:mongbi_app/features/auth/presentation/pages/widgets/mongbi_image_widget.dart';
+import 'package:mongbi_app/features/auth/presentation/pages/widgets/naver_login_button_widget.dart';
+import 'package:mongbi_app/features/auth/presentation/pages/widgets/text_widgets.dart';
 import 'package:mongbi_app/presentation/common/custom_snack_bar.dart';
 import 'package:mongbi_app/presentation/terms/widgets/terms_bottom_sheet_layout_widget.dart';
 import 'package:mongbi_app/providers/auth_provider.dart';
