+++ lib/features/auth/presentation/view_models/auth_view_model.dart
@@ -6,10 +6,10 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:kakao_flutter_sdk_user/kakao_flutter_sdk_user.dart' as kakao;
 import 'package:mongbi_app/core/exceptions/auth_custom_exception.dart';
-import 'package:mongbi_app/domain/entities/user.dart';
-import 'package:mongbi_app/domain/use_cases/login_with_apple.dart';
-import 'package:mongbi_app/domain/use_cases/login_with_kakao.dart';
-import 'package:mongbi_app/domain/use_cases/login_with_naver.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
+import 'package:mongbi_app/features/auth/domain/use_cases/login_with_apple.dart';
+import 'package:mongbi_app/features/auth/domain/use_cases/login_with_kakao.dart';
+import 'package:mongbi_app/features/auth/domain/use_cases/login_with_naver.dart';
 import 'package:mongbi_app/providers/account_provider.dart';
 import 'package:mongbi_app/providers/auth_provider.dart';
 import 'package:mongbi_app/providers/user_info_provider.dart';
