+++ lib/providers/account_provider.dart
@@ -1,9 +1,9 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/account_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_account_data_source.dart';
-import 'package:mongbi_app/data/repositories/remote_account_repository.dart';
-import 'package:mongbi_app/domain/repositories/account_repository.dart';
 import 'package:mongbi_app/domain/use_cases/remove_account_use_case.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/account_data_source.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_account_data_source.dart';
+import 'package:mongbi_app/features/auth/data/dtos/remote_account_repository.dart';
+import 'package:mongbi_app/features/auth/domain/repositories/account_repository.dart';
 import 'package:mongbi_app/providers/core_providers.dart';
 
 final _accountDataSourceProvider = Provider<AccountDataSource>((ref) {
