+++ lib/providers/auth_provider.dart
@@ -1,14 +1,14 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:mongbi_app/data/data_sources/remote_apple_auth_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_kakao_auth_data_source.dart';
-import 'package:mongbi_app/data/data_sources/remote_naver_auth_data_source.dart';
-import 'package:mongbi_app/data/repositories/remote_auth_repository.dart';
-import 'package:mongbi_app/domain/entities/user.dart';
 import 'package:mongbi_app/domain/repositories/auth_repository.dart';
-import 'package:mongbi_app/domain/use_cases/login_with_apple.dart';
-import 'package:mongbi_app/domain/use_cases/login_with_kakao.dart';
-import 'package:mongbi_app/domain/use_cases/login_with_naver.dart';
-import 'package:mongbi_app/presentation/auth/viewmodels/auth_view_model.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_apple_auth_data_source.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_kakao_auth_data_source.dart';
+import 'package:mongbi_app/features/auth/data/data_sources/remote_naver_auth_data_source.dart';
+import 'package:mongbi_app/features/auth/data/dtos/remote_auth_repository.dart';
+import 'package:mongbi_app/features/auth/domain/entities/user.dart';
+import 'package:mongbi_app/features/auth/domain/use_cases/login_with_apple.dart';
+import 'package:mongbi_app/features/auth/domain/use_cases/login_with_kakao.dart';
+import 'package:mongbi_app/features/auth/domain/use_cases/login_with_naver.dart';
+import 'package:mongbi_app/features/auth/presentation/view_models/auth_view_model.dart';
 import 'package:mongbi_app/providers/core_providers.dart';
 
 final _remoteNaverAuthDataSourceProvider = Provider<RemoteNaverAuthDataSource>(
